provide the correct place for multiTenantFilter in web.xml
Previously, the plugin puts the multiTenantFilter as the very first filter. This leads to trouble if spring-security-core is used in the project since spring-security-core'
s filter is add to a later point in the chain, a TenantResolver could not access SpringSecurityContextHolder. So we must make sure that tenantFilter is processed after the spring-security-core filter. SSC hooks its filter right after grailsWebRequest and does this in _Events as post operation - for multiTenant it is sufficient to place its filter after grailsWebRequest
So the order is now:
charEncodingFilter - hiddenHttpMethod - grailsWebRequest - springsecuritycore - mulitTenant - sitemesh